refactor: create fhe related module move encrypt_string and decrypt_chunks to fhe.rs
