+++ src/utils/base64.rs
@@ -1,6 +1,5 @@
 extern crate base64;
-
-use base64::{engine::general_purpose::STANDARD, Engine as _};
+use base64::{engine::general_purpose::STANDARD, DecodeError, Engine as _};
 
 pub fn encode(input: &str) -> String {
     return STANDARD.encode(input.as_bytes());
@@ -15,5 +14,25 @@ pub fn decode(input: &str) -> String {
         Err(err) => {
             return format!("Decoding error: {:?}", err);
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    #[test]
+    fn encode() {
+        let input: &str = "Hello, World!";
+        let expected: &str = "SGVsbG8sIFdvcmxkIQ==";
+        let result: String = STANDARD.encode(input);
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn decode() {
+        let input: &str = "SGVsbG8sIFdvcmxkIQ==";
+        let expected: &str = "Hello, World!";
+        let result: Result<Vec<u8>, DecodeError> = STANDARD.decode(input);
+        assert_eq!(result.unwrap(), expected.as_bytes());
     }
 }
